Differentiate tests for 2.12, 2.13 and 3+ scala versions
