+++ modules/integration/src/test/scala/scala/cli/integration/ScalafixTests212.scala
@@ -1,0 +1,5 @@
+package scala.cli.integration
+
+class ScalafixTests212 extends ScalafixTestDefinitions with Test212 {
+  override val unusedRuleOption: String = "-Ywarn-unused"
+}
