+++ modules/integration/src/test/scala/scala/cli/integration/ScalafixTests3.scala
@@ -1,0 +1,5 @@
+package scala.cli.integration
+
+class ScalafixTests3 extends ScalafixTestDefinitions with TestDefault {
+  override val unusedRuleOption: String = "-Wunused:all"
+}
